+++ uiTestGlaze/src/main/java/com/getyourguide/uitestglazesample/UiTestGlaze.kt
@@ -20,14 +20,14 @@ data class UiTestGlaze(
     private val device = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
 
     fun tap(
-        uiElement: UiElement,
+        uiElementIdentifier: UiElementIdentifier,
         optional: Boolean = false,
         retryCount: Int = 3,
         longPress: Boolean = false
     ) {
         val hierarchy =
             HierarchySettleHelper.waitTillHierarchySettles(config.loadingResourceIds, device)
-        tapHelper.tap(uiElement, optional, retryCount, longPress, hierarchy, device)
+        tapHelper.tap(uiElementIdentifier, optional, retryCount, longPress, hierarchy, device)
     }
 
     fun scroll(scrollOption: ScrollOption) {
@@ -41,14 +41,14 @@ data class UiTestGlaze(
         return AssertionHelper.assert(assertion, optional, hierarchy, device)
     }
 
-    fun find(uiElement: UiElement): FoundUiElement? {
+    fun find(uiElementIdentifier: UiElementIdentifier): UiElement? {
         val hierarchy =
             HierarchySettleHelper.waitTillHierarchySettles(config.loadingResourceIds, device)
-        return FindUiElementHelper.getUiElement(uiElement, hierarchy, true, device)
+        return FindUiElementHelper.getUiElement(uiElementIdentifier, hierarchy, true, device)
     }
 
-    fun inputText(text: String, uiElement: UiElement) {
-        InputTextHelper.inputText(text, uiElement, device)
+    fun inputText(text: String, uiElementIdentifier: UiElementIdentifier) {
+        InputTextHelper.inputText(text, uiElementIdentifier, device)
         HierarchySettleHelper.waitTillHierarchySettles(config.loadingResourceIds, device)
     }
 
@@ -66,47 +66,55 @@ data class UiTestGlaze(
 
 }
 
-sealed class UiElement(open val index: Int = 0) {
-    data class Id(@IdRes val id: Int, override val index: Int = 0) : UiElement(index)
-    data class TestTag(val testTag: String, override val index: Int = 0) : UiElement(index)
+sealed class UiElementIdentifier(open val index: Int = 0) {
+    data class Id(@IdRes val id: Int, override val index: Int = 0) : UiElementIdentifier(index)
+    data class TestTag(val testTag: String, override val index: Int = 0) :
+        UiElementIdentifier(index)
+
     data class Text(
         val text: String,
         val ignoreCase: Boolean = false,
         override val index: Int = 0
-    ) : UiElement(index)
+    ) : UiElementIdentifier(index)
 
     data class TextResource(
         @StringRes val stringResourceId: Int,
         override val index: Int = 0
-    ) : UiElement(index)
+    ) : UiElementIdentifier(index)
 
-    data class TextRegex(val textRegex: Regex, override val index: Int = 0) : UiElement(index)
+    data class TextRegex(val textRegex: Regex, override val index: Int = 0) :
+        UiElementIdentifier(index)
+
     data class ChildFrom(
-        val uiElementParent: UiElement,
-        val uiElementChild: UiElement,
+        val uiElementIdentifierParent: UiElementIdentifier,
+        val uiElementIdentifierChild: UiElementIdentifier,
         val inputIndicatorText: Boolean = true
-    ) :
-        UiElement()
+    ) : UiElementIdentifier()
+
+    data class PositionInHierarchy(
+        override val index: Int = 0,
+        val inputIndicatorText: Boolean = true
+    ) : UiElementIdentifier(index)
 }
 
 sealed interface ScrollOption {
-    data class VerticalDown(val inUiElementId: UiElement) : ScrollOption
-    data class HorizontalRight(val inUiElementId: UiElement) : ScrollOption
+    data class VerticalDown(val inUiElement: UiElementIdentifier) : ScrollOption
+    data class HorizontalRight(val inUiElement: UiElementIdentifier) : ScrollOption
     data class Manual(val startX: Int, val startY: Int, val endX: Int, val endY: Int) : ScrollOption
     data class VerticalDownToElement(
-        val toUiElement: UiElement,
-        val inUiElementId: UiElement
+        val toUiElement: UiElementIdentifier,
+        val inUiElement: UiElementIdentifier
     ) : ScrollOption
 
     data class HorizontalRightToElement(
-        val toUiElement: UiElement,
-        val inUiElementId: UiElement
+        val toUiElement: UiElementIdentifier,
+        val inUiElement: UiElementIdentifier
     ) : ScrollOption
 }
 
 sealed interface Assertion {
-    data class Visible(val uiElement: UiElement) : Assertion
-    data class NotVisible(val uiElement: UiElement) : Assertion
+    data class Visible(val uiElementIdentifier: UiElementIdentifier) : Assertion
+    data class NotVisible(val uiElementIdentifier: UiElementIdentifier) : Assertion
 }
 
 sealed interface PressKey {
@@ -119,7 +127,7 @@ sealed interface PressKey {
     object VolumeDown : PressKey
 }
 
-data class FoundUiElement(
+data class UiElement(
     val x: Int,
     val y: Int,
     val width: Int,
@@ -129,5 +137,5 @@ data class FoundUiElement(
     val clickable: Boolean? = null,
     val checked: Boolean? = null,
     val enabled: Boolean? = null,
-    val treeNode: TreeNode? = null
+    val children: List<UiElement>? = null
 )